+++ src/views/components/messageTypes/VideoMessage.js
@@ -3,7 +3,7 @@ import { Text, View } from 'react-native';
 import users from '../../../services/user';
 import { useObservableState } from 'observable-hooks';
 import { BubbleWrapper, SenderText } from '../MessageItem';
-import Video from 'react-native-video';
+import { WebView } from 'react-native-webview';
 
 export default function VideoMessage({
   message,
@@ -20,9 +20,20 @@ export default function VideoMessage({
   const status = useObservableState(message.status$);
   const isMe = myUser.id === message.sender.id;
 
+  if (!content) return null;
+
   const _onLongPress = () => onLongPress(message);
   const _onPress = () => onPress(message);
   const _onSwipe = () => onSwipe(message);
+
+  const { width, height } = content.thumb;
+
+  const iframeHtml = `<iframe src="${content.url}" frameborder="0" width="${width}" height="${height}" style="height:${height}px;width:${width}px;"></iframe>`;
+  const videoHtml = `<video src="${content.url}" alt="science.mp4" controls="" preload="none" autoplay="" playsinline height="100%" width="100%" style="background-color:#ddd" ></video>`;
+  const head = `<style>body{margin:0}</style><meta name="viewport" content="width=device-width, initial-scale=1">`;
+  const html = `<!DOCTYPE html><html><head>${head}</head><body>${iframeHtml}</body></html>`;
+
+  const test = `<html style="background-color:#ddd"><head><meta name="viewport" content="width=${width}"></head><body><video controls="true" autoplay="true" name="media" width="${width}"><source src="${content.url}" type="video/mp4"></video></body></html>`;
 
   return (
     <>
@@ -32,13 +43,29 @@ export default function VideoMessage({
         onSwipe={onSwipe ? _onSwipe : null}
         receipts={receipts}>
         <View style={{ borderRadius: 20, overflow: 'hidden' }}>
-          <Video
-            source={{ uri: 'http://techslides.com/demos/sample-videos/small.mp4' }}
-            //  onBuffer={this.onBuffer}                // Callback when remote video is buffering
-            //  onError={this.videoError}
-            style={{ width: 300, height: 200 }}
-            controls
-            pictureInPicture
+          <WebView
+            source={{ html: test }}
+            style={{
+              width,
+              height,
+              backgroundColor: '#ddd',
+            }}
+            onFileDownload={({ nativeEvent: { downloadUrl } }) => {
+              console.log('download ', downloadUrl);
+              // You use downloadUrl which is a string to download files however you want.
+            }}
+            onError={console.warn}
+            allowsFullscreenVideo
+            useWebKit
+            javascriptEnabled
+            originWhitelist={['*']}
+            bounces={false}
+            dataDetectorTypes="link"
+            scalesPageToFit={false}
+            scrollEnabled={false}
+            automaticallyAdjustContentInsets={false}
+            allowsInlineMediaPlayback
+            mediaPlaybackRequiresUserAction={false}
           />
         </View>
       </BubbleWrapper>